bad wip walk cycle (read desc)
I am committing this not because this walk cycle is anywhere near done, in fact the only thing "done" is the up-down motion, but because this has been sitting open on my computer for almost 2 weeks now and priorities have long shifted away from this and towards gameplay. I fear that eventually my pc will restart or something and I lose this lmao.